pyproject.toml
[project]
name = "next_web_ai_backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Robin.Wang", email = "3Jg$94sbQ@q1"},
]
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic-settings>=2.2.1",
    "boto3>=1.34.102",
    "logtail-python>=0.2.10",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
start = "doppler run -- uvicorn app.main:app --reload"
